+++ covalent_ui/api/v1/routes/end_points/electron_routes.py
@@ -27,6 +27,7 @@ from sqlalchemy.orm import Session
 
 import covalent_ui.api.v1.database.config.db as db
 from covalent._results_manager.results_manager import get_result
+from covalent_dispatcher._core.execution import _get_task_inputs as get_task_inputs
 from covalent_ui.api.v1.data_layer.electron_dal import Electrons
 from covalent_ui.api.v1.models.electrons_model import (
     ElectronExecutorResponse,
@@ -87,7 +88,6 @@ def get_electron_inputs(dispatch_id: uuid.UUID, electron_id: int) -> str:
     Returns:
         Returns the inputs data from Result object
     """
-    from covalent_dispatcher._core.execution import _get_task_inputs as get_task_inputs
 
     result_object = get_result(dispatch_id=str(dispatch_id), wait=False)