+++ tests/covalent_ui_backend_tests/end_points/electrons_test.py
@@ -273,8 +273,9 @@ def test_electrons_file_bad_request():
         path=test_data["path"],
     )
     assert response.status_code == test_data["status_code"]
-    response_detail = response.json()["detail"][0]
-    assert response_detail["type"] == "type_error.enum"
+    if "response_message" in test_data:
+        response_detail = response.json()["detail"][0]
+        assert response_detail["msg"] == test_data["response_message"]
 
 
 def test_electrons_inputs_bad_request():